requirements.txt
opencv-python
mediapipe
numpy
google-generativeai
cvzone